Zip Methode im Repository Manager
