+++ CMI/Manager/Repository/CMI.Manager.Repository.csproj
@@ -169,7 +169,6 @@
     <Compile Include="RepositoryManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RepositoryService.cs" />
-    <Compile Include="Systems\Rosetta\RosettaZipFileManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\Contract\Common\Gebrauchskopie\ablieferung.xsd">
